requirements.txt
Flask==3.0.0
line-bot-sdk>=3.0.0
requests==2.31.0
openai>=1.3.0
httpx==0.27.2
gspread==6.1.2
google-auth==2.29.0
pyswisseph==2.10.3.2
